make gitpod work with git 2.18+ (see comments in code for details)
Due to a failure in running it under git 2.18, I had no choice but to
replace the entire thing with one shell script!

documentation is within the script, including details of said failure.

if you need any of the hardcodings fixed, please open an issue and I
will work on it.